Rename library
Yeah... ALfoNS just doesn't really resonate with me anymore; besides, having to append the ".js" for npm sucks. But in a sense, this is just returning to what the name originally-originally was - in my "owl-old" repo for the OWL Bot, I originally included a built-in NationStates API library called "nationscripts", which this here library is based on.