Add missing enum values
Found a missing Notice type - 'PW' for reports of bad login attempts - and two missing census scale IDs - for the region-specific alphabetical order and number-of-nations scales - and added both accordingly.

Furthermore, the injunction badges are now live on-site and accordingly visible via the API, so the injunction WABadgeType has been uncommented.