Fix JSDocs... again
Apparently my module headers in the main files were confusing VSCode IntelliSense when using the package externally, so yeah, they've been nuked and replaced by the original Mozilla licence headers suggested for the MPL again.

Also, added a bunch of ignore tags to internal parser functions etc. that aren't intended to show up when generating the documentation.

Finally, as a cosmetic change, the name for the documentation Action is now a lowercase "docs", so the badge for it better fits in with the others in the readme :D