Fix JSDocs, for real this time
Turns out that correlation does, in fact, not equal causation, and my previous theory was false. Instead it has... something...? to do with when/where the modules are require()'d? I have literally no idea why, but when copying all the request type require()s into the index file as well, they seem to be loaded fine.

Might also just be a VS Code thing, since I'm pretty sure that in the original version 1.0.0, the typings worked when I tested it. ¯\_(ツ)_/¯