Add NoticeIcon enum
Added a NoticeIcon enum to the enums module in order to complement the NoticeType enum, since the values for the notice icons returned by the API can have a couple weird names.

Re-organized the NoticeType enum to be ordered alphabetically.

Added the `RMB_QUOTE_DISPATCH` NoticeType.

(Version number preemptively bumped.)