Create basic factories
Since the new XML parsing will translate the XML directly into NationScript objects - without the intermediate step of creating and then reading from a generic JS object in between - there needs to be some way of handling the XML parsing events.

The newly introduced factory class structure should hopefully allow to streamline this in a way similar to the current "parseX" functions of the typedefs folder, using nested lines of ever more specialised sub-factories. The ones created now are very abstract versions for parsing to an object and an array respectively.