Move factory code
The file containing the base factory code has been moved from `/src/factories/base.js` to `/src/factory.js`.

The plan when starting the factory class was to use an inheritance-based system similar to the request classes, so it was put in its own folder, but this has long been changed to the configuration approach using `.onCase()` on factory instances. As such, a folder just for factories isn't necessary anymore.