Automatic root detection
Removed the reliance of Factories on a hard-defined root node; JSDocs updated accordingly.

A factory receiving a newly opened tag for handling will now *require* a sub-factory to be assigned for handling the content of that tag. While this means that a factory will assign itself a sub-factory simply for reading the text content of a node, this removes the need for a defined root node name, since receiving an event for a closed tag without having a sub-factory assigned automatically means that this was the receiving factory's root node.