Fix DumpCard-related factories
A misconfiguration of the DumpCard factory has been fixed. In the sub-factory for CardBadges, the direct parsing in the ArrayFactory produced duplicate SC badges, which the extraction into a dedicated NSFactory fixed. Relatedly, the missing CardBadgeType for forum-only moderators has been added. The Trophy factory was missing a translation case for the `AVERAGE_INCOME` scale, which has been added.