Update JSDocs
The JSDocs for the whole project have been generally expanded and clarified, and particularly adjusted to better follow JSDoc's namepaths. Since JSDoc and VSCode annoyingly work differently regarding namepaths, this is only possible within limited confines so as to not break the IntelliSense, sadly. The `jsdoc-tsimport-plugin` somewhat alleviates this, but still.