Add option warnings
NationStates has added a `<WARNING>` tag to the data returned by the `issues` shard of the Nation API. This adds a corresponding `warnings` property to the `Issue` type to hold that data, and updates the factory configuration accordingly. Updated the sample XML to incorporate an instance of a warning message.

Version number bumped for a new minor release of the package.